// ui/src/generated/TrackerProjectResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BoatImportReport } from "./BoatImportReport";
import type { ProjectSummary } from "./ProjectSummary";

export type TrackerProjectResult = { project: ProjectSummary, boats: Array<BoatImportReport>, warnings: Array<string>, };